Update README.md
Netflix is one of the most well-known media and video streaming services. They offer more than 10000 movies and TV episodes on their platform, and as of the middle of 2023, they had more than 230 million subscribers worldwide. Because of this, Netflix is a popular streaming service with a wide range of users.

However, if subscribers to Netflix begin binge-watching a particular genre of films or television shows, they are likely to look for more of the same to watch. How could customers sort through more than 10 million films to truly locate the ones they like? They won't likely do that because it takes a lot of time.
Thus, the significance of clustering.

By prioritizing the movies or series that fall under the same genre as the movies that users typically watch, Netflix is able to provide its users with satisfactory results.Customers will save a ton of time, and Netflix will make a ton of money.

In this unsupervised learning project, we are going to identify similar movies or series based on movie characteristics with the help of a suitable data set. This tabular dataset consists of listings of the movies and tv shows available on Netflix, along with details such as  cast, directors, ratings, release year, duration and so on.

The objective of this research is to categorize or group the Netflix shows into distinct clusters in which the shows are similar to one another and different to one another.
 
 We will thoroughly cover each of the following steps during this project:
 
 1- Clean, explore and visualize the dataset.
 
 2- Perform clustering algorithm techniques, after which we will be able to understand how different and similar shows are from one another
 
 3- Use these clusters to provide customers with tailored show recommendations based on their tastes.


Below is the link to access the dataset:

 [netflix titles dataset](https://www.kaggle.com/datasets/shivamb/netflix-shows/download?datasetVersionNumber=5)